COMP: fixed warnings

+++ VTK/Rendering/vtkGLSLShaderProgram.cxx
@@ -152,7 +152,7 @@ void vtkGLSLShaderProgram::GetProgramInfo()
   GLint type = 0;
   vtkgl::GetProgramiv( static_cast<GLuint>(this->Program), 
     vtkgl::OBJECT_TYPE_ARB, &type);
-  if( type == vtkgl::PROGRAM_OBJECT_ARB )
+  if( type == static_cast<GLint>(vtkgl::PROGRAM_OBJECT_ARB) )
     {
     infoString += "GLSL Program. \n";
     }
